requirements.txt
flask==2.3.3
pytesseract==0.3.10
opencv-python==4.8.0.76
spacy==3.6.1
python-dotenv==1.0.0
pymongo==4.5.0
psycopg2-binary==2.9.7
Pillow==10.0.0
numpy==1.24.3
python-magic==0.4.27
pdf2image==1.16.3
flask-cors==4.0.0
gunicorn==21.2.0 